+++ packages/cookie/src/index.ts
@@ -1,1 +1,1 @@
-export { type CookieOptions, Cookie } from './lib/cookie.ts'
+export { createCookie, type CookieOptions, Cookie } from './lib/cookie.ts'