+++ packages/cookie/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This is the changelog for [`cookie`](https://github.com/remix-run/remix/tree/main/packages/cookie). It follows [semantic versioning](https://semver.org/).
 
+## v0.3.0 (2025-11-08)
+
+- BREAKING CHANGE: Rename `cookie.isSigned` to `cookie.signed`
+- Add `createCookie` function to create a new `Cookie` object
+- `CookieOptions` now extends `CookieProperties` so all cookie properties may be set in the `Cookie` constructor
+
 ## v0.2.0 (2025-11-04)
 
 - Update `@remix-run/headers` peer dep to v0.15.0